+++ study/src/main/java/com/myzuji/study/easyexcel/read/DemoData0.java
@@ -1,14 +1,16 @@
-package com.myzuji.study.java.easyexcel.read;
+package com.myzuji.study.easyexcel.read;
 
 import com.alibaba.excel.annotation.format.DateTimeFormat;
-import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.util.Date;
 
 /**
  * 基础数据类.这里的排序和excel里面的排序一致
  **/
-@Data
+@Getter
+@Setter
 public class DemoData0 {
     private String orderNo;
     @DateTimeFormat("yyyyMMddHHmmss")
