+++ study/src/main/java/com/myzuji/study/easyexcel/read/DemoData1.java
@@ -1,9 +1,11 @@
-package com.myzuji.study.java.easyexcel.read;
+package com.myzuji.study.easyexcel.read;
 
 import com.alibaba.excel.annotation.ExcelProperty;
-import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 public class DemoData1 {
 
     @ExcelProperty("商户订单号")
