+++ study/src/main/java/com/myzuji/study/spi/SPITest.java
@@ -1,9 +1,7 @@
 package com.myzuji.study.spi;
 
-import sun.misc.Service;
-
-import java.util.Iterator;
-import java.util.ServiceLoader;
+import jdk.nashorn.internal.ir.annotations.Ignore;
+import org.junit.jupiter.api.Test;
 
 /**
  * 说明
@@ -13,18 +11,20 @@ import java.util.ServiceLoader;
  */
 public class SPITest {
 
-    public static void main(String[] args) {
-        Iterator<SPIService> providers = Service.providers(SPIService.class);
-        while (providers.hasNext()) {
-            SPIService spiService = providers.next();
-            spiService.execute();
-        }
-        System.out.println("============");
-        ServiceLoader<SPIService> loader = ServiceLoader.load(SPIService.class);
-        Iterator<SPIService> spiServiceIterator = loader.iterator();
-        while (spiServiceIterator.hasNext()) {
-            SPIService spiService = spiServiceIterator.next();
-            spiService.execute();
-        }
+    @Test
+    @Ignore
+    void sPITest() {
+//        Iterator<SPIService> providers = Service.providers(SPIService.class);
+//        while (providers.hasNext()) {
+//            SPIService spiService = providers.next();
+//            spiService.execute();
+//        }
+//        System.out.println("============");
+//        ServiceLoader<SPIService> loader = ServiceLoader.load(SPIService.class);
+//        Iterator<SPIService> spiServiceIterator = loader.iterator();
+//        while (spiServiceIterator.hasNext()) {
+//            SPIService spiService = spiServiceIterator.next();
+//            spiService.execute();
+//        }
     }
 }
